// web-application/src/types-openapi/api/models/CreateInventoryCheckRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { CreateInventoryCheckItem } from './CreateInventoryCheckItem';
import {
    CreateInventoryCheckItemFromJSON,
    CreateInventoryCheckItemFromJSONTyped,
    CreateInventoryCheckItemToJSON,
    CreateInventoryCheckItemToJSONTyped,
} from './CreateInventoryCheckItem';

/**
 * 
 * @export
 * @interface CreateInventoryCheckRequest
 */
export interface CreateInventoryCheckRequest {
    /**
     * 
     * @type {Date}
     * @memberof CreateInventoryCheckRequest
     */
    checkTime?: Date;
    /**
     * 
     * @type {string}
     * @memberof CreateInventoryCheckRequest
     */
    note?: string | null;
    /**
     * 
     * @type {Array<CreateInventoryCheckItem>}
     * @memberof CreateInventoryCheckRequest
     */
    items?: Array<CreateInventoryCheckItem> | null;
}

/**
 * Check if a given object implements the CreateInventoryCheckRequest interface.
 */
export function instanceOfCreateInventoryCheckRequest(value: object): value is CreateInventoryCheckRequest {
    return true;
}

export function CreateInventoryCheckRequestFromJSON(json: any): CreateInventoryCheckRequest {
    return CreateInventoryCheckRequestFromJSONTyped(json, false);
}

export function CreateInventoryCheckRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateInventoryCheckRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'checkTime': json['checkTime'] == null ? undefined : (new Date(json['checkTime'])),
        'note': json['note'] == null ? undefined : json['note'],
        'items': json['items'] == null ? undefined : ((json['items'] as Array<any>).map(CreateInventoryCheckItemFromJSON)),
    };
}

export function CreateInventoryCheckRequestToJSON(json: any): CreateInventoryCheckRequest {
    return CreateInventoryCheckRequestToJSONTyped(json, false);
}

export function CreateInventoryCheckRequestToJSONTyped(value?: CreateInventoryCheckRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'checkTime': value['checkTime'] == null ? undefined : ((value['checkTime']).toISOString()),
        'note': value['note'],
        'items': value['items'] == null ? undefined : ((value['items'] as Array<any>).map(CreateInventoryCheckItemToJSON)),
    };
}